v1.0 GammaRay
Gamma Tool/Game Element introduced

>>>>>>>>NEW<<<<<<<<<<<
--------- GammaRay Prefab -------
>GammaRayController.cs
Controls the overall tool elements.
Holds the pivot information to randomly alter the speed of the rotation of the generators

>GammaGenerator.cs
Creates and shoots "photon rays" elements at asteroid
Randomly ejects an emitterRay object towards the asteroid
The direction of the ray is slightly varied (but always at the object)

>NeutronEmitter.cs
Emits a random number of neutron elements on collision with asteroid
When this "photo ray" element collides with the asteroid it destroys itself
Randomly creates a number of neutron elements to replace it
generated neutrons are given an ID (type) and velocity relative to its collision angle/speed

>Neutron.cs
The actual score/point element
generated by the NeutronEmitter
Sets a timer for its own self-destruction after a given time (prevent object overload)
Handles the collision into player methods for score keeping.
Destroys itself upon collection

----------BGSpaceTrigger.cs------------
Added a "play field" trigger for the game.
The BG sphere now has an interior 2D collider defining the edge of the play area
If the player tagged object exits this collider, it deactivates the player
Basically if the player is this far out, they have exceeded the gravity of the asteroid and failed
Can be build upon later with more complex "game over" methods.

>>>>>>>>CHANGES<<<<<<<<<<<<

--------ObjectRotate.cs--------
Changed rotation speed to be a preset speed, but alterable
Added Set methods for both the rotation axis and speed of rotation

--------Constants.cs--------
Added constants information for the GammaRay tools for easier accessiblity and handling for later gameplay mechanic balancing.

---------GameRoot.cs---------
Added a point scoring system for the Neutron (gamma ray) collection