Near Final Build
Stage Changing integration done.  Player can start at Level1, and play through to the Credits.  Inbetween each stage it goes to EndOfLevel, which is the DSOC trivia questions.

There can be ONLY ONE GameRoot/GameController.  The stage you press start in should have the GameController game object enabled.  It should be turned off in any and all other scenes.  The GameController is left in those scenes (deactivated) if needed to debug in that scene.