WebGL fixes and HUD scaling Fixes
Hopefully this fixes the newest WebGL crash.  The video player had a problem when fading because the volume level was linked to the Alpha channel of the screen.  Which makes sense, but if the alpha goes negative it causes the volume to go negative which crashed the player.  This puts a check in that the volume being sent is above 0.001

Also, made some changes to the ingame HUD.  The hud should now scale appropriately when changing to fullscreen.