VidPlayer Changes
Trying something different with the vid player.  It's a bit hacky, but maybe it works for the deadline.  Instead of trying to have the code find which video to play, buffer that video and play to the same Output Render Texture, I created RenderTextures for each video type.  Also, the Level controller has the video pre linked in Unity Inspector, rather than trying to find it.  This should mean that it buffers the video on load, rather than after loading the code.  We'll see if it works after it pushes to WebGL.

I will say that this change has fixed the "videoEnd" event invokes.  That's working properly now.