Crash Fixes, Removal of Allin1Sprite, Vid Fixes
Changed the video player files from being auto-play, to being called on Update().  This seems to fix the WebGL problem since the videos aren't being played on Awake() and getting blocked by the browser's ad-block.
Changed the video outputs so that each video clip has it's own render texture (I don't know if this helps, I just thought maybe it would fix some bugs and maybe prevent the screen lag as the video switches from one to another.

Removed the Allin1SpriteShader.
This was a bit of work rooting out the plugin.  It worked great for playing on desktop, but something about these shaders causes crashes on WebGL.  So I found all instances where these materials and shaders were being used and removed them.  This mostly affected the Magnetometer tool.  Since, most of the mag tool flux lines are rendered with this sprite shader.  But I made some changes to the tool and added some new variablitily to the flux line generation that almost makes up for it.

Lastly, found the bug that literally took all my time and it was because Unity and WebGL are stupid.  The DSOC questions kept crashing on load, but only on WebGL.  said there was a null object.  Couldn't figure it out.  After 10 webgl rebuilds, I narrowed it down to the UIText object for the DragNDrop items.  Apparently, it wasn't find the TextMeshProUI text on the webgl.  I replaced the "FindComponentInChildren" with just specifically referencing the object in the inspector.  My guess is that the multiple "findcomponentinchildren" calls just weren't jiving in JavaScript.