F4.26 Radio Ring Performance Pass
Fixed a bug in the Radio Ring generation that drawing the rings by Degrees rather than Radians.
The result was a several Line Renderers with more than 53k positions.  This fix reduces that to 100-300 points per ring (depending on size).
Also reduced the amount of actions taking place in Radio Rings Update methods.
Update now only adds points to the score and is not responsible for playing the audio or changing the pitch.